+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Change Log
 - Change to save `com.apple.TextEncoding` xattr only if the file already has the encoding xattr or it's a new document.
 - Move removed themes/styles to Trash instead delete them immediately.
 - Now, Utility actions perform with multiple selection.
+- Avoid showing not-writable alert on Resume again.
 - Delay timing to save text key bindings setting.
 - Localize document types.
 - Improve text rendering with non-opaque view.